Move example project to subfolders so we can test that files in subfolders are correctly reported in the html reporter
