+++ integrationtests/TargetProjects/NetCore2_1/ExampleProject.XUnit/DummyMathTest.cs
@@ -1,3 +1,4 @@
+using ExampleProject.Math;
 using Xunit;
 
 namespace ExampleProject.XUnit
