+++ integrationtests/TargetProjects/NetCore2_1/ExampleProject/Math/Calc/DummyCalc.cs
@@ -1,18 +1,20 @@
 ﻿using ExampleClassLibrary;
 
-namespace ExampleProject
+namespace ExampleProject.Math.Calc
 {
     public class DummyCalc
     {
         public int SomeCalc(int first, int second)
         {
-			while(second > 0) {
-				first = first + second + 1;
-				second--;
-			}
-			while(1 < 0) {
-				// endless looping
-			}
+            while (second > 0)
+            {
+                first = first + second + 1;
+                second--;
+            }
+            while (1 < 0)
+            {
+                // endless looping
+            }
             return first;
         }
