+++ integrationtests/TargetProjects/NetCore2_1/ExampleProject/Math/DummyMath.cs
@@ -1,4 +1,4 @@
-﻿namespace ExampleProject
+﻿namespace ExampleProject.Math
 {
     public class DummyMath
     {
